% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{make_curl_req}
\alias{make_curl_req}
\title{Make internal request to GitHub API}
\usage{
make_curl_req(..., endpoint = "repos", owner = "insightsengineering")
}
\arguments{
\item{...}{path in API}

\item{endpoint}{endpint in API}

\item{owner}{organization on owner}
}
\value{
a httr2 query
}
\description{
Make internal request to GitHub API
}
\examples{
make_curl_req("repos", "insightsengineering", "teal", "actions", "workflows", "scheduled.yaml", "runs")
}
