% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_pkg_table}
\alias{build_pkg_table}
\title{Build package table information on jobs}
\usage{
build_pkg_table(
  pkg_name,
  yaml_files = c("scheduled.yaml", "scheduled.yml"),
  workflows = c("min_cohort", "min_isolated", "max", "release"),
  keep_regex = "Dependency Test - (.*) 🔢 .*"
)
}
\arguments{
\item{pkg_name}{name of package}

\item{yaml_files}{yaml files to query}

\item{workflows}{job workflows to query (placeholder on failure)}

\item{keep_regex}{regex to keep on jobs}
}
\value{
a tibble with package, name, status, url and icon
}
\description{
Build package table information on jobs
}
\examples{
build_pkg_table("teal")
build_pkg_table(
  "teal",
  workflows = c("windows (r-devel-windows-x86_64)","macos-arm64 (r-release-macos-arm64)","ubuntu-clang","macos (r-release-macos-x86_64)","ubuntu-gcc12","windows (r-release-windows-x86_64)","macos-arm64 (r-oldrel-macos-arm64)","ubuntu-next","macos (r-oldrel-macos-x86_64)","ubuntu-release","windows (r-oldrel-windows-x86_64)","linux","nosuggests"),
  keep_regex = "R-hub 🌐 / (.*) \\\\(.*\\\\)"
)
}
